correcciones completadas

+++ src/routes/carts.routes.js
@@ -1,5 +1,6 @@
 import { json, Router } from "express";
 import { CartManager } from "../dao/cartManager.js";
+import { ProductManager } from "../dao/productManager.js"
 import { isValidObjectId } from "mongoose";
 
 
@@ -7,8 +8,8 @@ export const cartsRouter = Router();
 
 cartsRouter.post('/', async (req, res) => {
     try {
-        await CartManager.createCart();
-        res.status(201).json({mesagge: "El cart fue creado con existo"});
+        const cart = await CartManager.createCart();
+        res.status(201).json({mesagge: "El cart fue creado con existo", cart});
         // return res.setHeader("Content-Type", "Application/json");
     } catch (error) {
         res.status(500).json({error: "No se pudo crear el carrito"})
@@ -43,12 +44,22 @@ cartsRouter.post('/:cid/products/:pid', async (req, res) => {
     }
     
     try {
+
+        const product = await ProductManager.getProductsById(pid);
+        if (!product) {
+            return res.status(404).json({error: "El producto no existe en la base de datos"})
+        }
+
+        const cartExist = await CartManager.getCart(cid);
+        if (!cartExist) return res.status(404).json({error: `El cart con id ${cid} no existe`});
+         
+
         await CartManager.addProductToCart(cid, pid);
         res.setHeader("Content-Type", "Application/json");
         return res.status(200).json({message: "Se agrego el producto correctamente"});
     } catch (error) {
         res.setHeader("Content-Type", "Application/json");
-        return res.status(500).json({error: "Hubo un error de servidor"})
+        return res.status(500).json({error: "Hubo un error de servidor"})  
     }
 });
 
@@ -63,12 +74,20 @@ cartsRouter.delete('/:cid/products/:pid', async (req, res) =>{
         return res.status(400).json({message: "Pid o Cid invalido"});
     }
     try {
+
+        const cartExist = await CartManager.getCart(cid);
+        if (!cartExist) return res.status(404).json({error: `El cart con id ${cid} no existe`});
+
         await CartManager.deleteProductFrom(cid, pid);
         res.setHeader("Content-Type", "Application/json");
         res.status(200).json({message: "Producto eliminado con exito"});   
     } catch (error) {
         console.log("Error al encontrar el producto");
         res.setHeader("Content-Type", "Application/json");
+        return res.status(500).json({
+            error: "Error inesperado en el servidor",
+            detalle: `${error.message}`,
+        });
     }
 });
 
@@ -82,12 +101,20 @@ cartsRouter.delete('/:cid', async (req, res) =>{
         return res.status(400).json({message: "Cid invalido"});
     }
     try {
+
+        const cartExist = await CartManager.getCart(cid);
+        if (!cartExist) return res.status(404).json({error: `El cart con id ${cid} no existe`});
+
         await CartManager.deleteAllProductsFromCart(cid);
         res.setHeader("Content-Type", "Application/json");
         res.status(200).json({message: "El carrito se vacio con exito"});   
     } catch (error) {
         console.log("Error al encontrar el carrito");
         res.setHeader("Content-Type", "Application/json");
+        return res.status(500).json({
+            error: "Error inesperado en el servidor",
+            detalle: `${error.message}`,
+        });
     }
 });
 
@@ -105,16 +132,34 @@ cartsRouter.put('/:cid', async (req, res) => {
     if (!products ) return res.status(400).json({error: "Se tiene que pasar un array de productos como parametro"});
 
     try {
+
+        for (const prod of products) {
+            if (!isValidObjectId(prod.product)){
+                return res.status(400).json({error: `El id del producto ${prod.product} no es valido`})
+            } 
+            const prodExist = await ProductManager.getProductsById(prod.product);
+            if (!prodExist) {
+                return res.status(404).json({error: `El producto con id ${prod.product} no existe`});
+            }
+        };
+
+        const cartExist = await CartManager.getCart(cid);
+        if (!cartExist) return res.status(404).json({error: `El cart con id ${cid} no existe`});
+
         await CartManager.updateCart(cid, products);
         res.setHeader("Content-Type", "Application/json");
         res.status(200).json({message: "El carrito fue actualizado con exito"});   
     } catch (error) {
         console.log("Error al encontrar el carrito");
         res.setHeader("Content-Type", "Application/json");
+        return res.status(500).json({
+            error: "Error inesperado en el servidor",
+            detalle: `${error.message}`,
+        });
     }
 });
 
-cartsRouter.put('/:cid/carts/:pid', async (req, res) => {
+cartsRouter.put('/:cid/products/:pid', async (req, res) => {
     const { cid, pid } = req.params;
     const { quantity } = req.body;
 
@@ -130,11 +175,24 @@ cartsRouter.put('/:cid/carts/:pid', async (req, res) => {
     if(isNaN(Number(quantity))) return res.status(400),json({error: "El cuantity tiene que ser un numero"});
 
     try {
+
+        const prodExist = await ProductManager.getProductsById(pid);
+        if (!prodExist) {
+            return res.status(404).json({error: `El producto con id ${pid} no existe`});
+        }
+
+        const cartExist = await CartManager.getCart(cid);
+        if (!cartExist) return res.status(404).json({error: `El cart con id ${cid} no existe`});
+
         await CartManager.updateQuantityProduct(cid, pid, quantity);
         res.setHeader("Content-Type", "Application/json");
         res.status(200).json({message: "El carrito fue actualizado con exito"});   
     } catch (error) {
         console.log("Error al encontrar el carrito");
         res.setHeader("Content-Type", "Application/json");
+        return res.status(500).json({
+            error: "Error inesperado en el servidor",
+            detalle: `${error.message}`,
+        });
     }
 });
